DEV: PostProcessMixin to aide with chat message processing
